+++ src/utils/utils.py
@@ -206,7 +206,7 @@ def to_gas_amount(gas_amount: int) -> gateway_pb2.GasAmount:
 
 def from_gas_amount(gas_amount: gateway_pb2.GasAmount) -> int:
     i: int = str(gas_amount.gas_amount)[::-1].find('.')
-    return float(gas_amount.gas_amount * pow(10, i) * pow(10, gas_amount.exponent-i))
+    return int(gas_amount.gas_amount * pow(10, i) * pow(10, gas_amount.exponent-i))
 """
 
 
@@ -215,7 +215,7 @@ def to_gas_amount(gas_amount: int) -> gateway_pb2.GasAmount:
 
 
 def from_gas_amount(gas_amount: gateway_pb2.GasAmount) -> int:
-    return int(float(gas_amount.n))
+    return int(gas_amount.n)
 
 
 def peers_id_iterator(ignore_network: str = None) -> Generator[str, None, None]:
